refactor: Avoid unused variables
