+++ src/cli/parsers/commitizen.js
@@ -1,7 +1,5 @@
 import minimist from 'minimist';
 
-import { isString, isArray } from '../../common/util';
-
 export {
   parse
 };
