+++ src/commitizen/commit.js
@@ -1,9 +1,8 @@
 import path from 'path';
 
-import dedent from 'dedent';
 import cacheDir from 'cachedir';
 import { ensureDir } from 'fs-extra';
-import { commit as gitCommit, log } from '../git';
+import { commit as gitCommit } from '../git';
 import * as cache from './cache';
 
 export default commit;
